Increased the max cap of battery.
Fixed many presets for DU, but for now the ones that are standarized per map are the ones en VEManager, this ones are more customized for DU (pitch dark or so.)

Tweaked some stuff related to Nightvision presets.

Attempt to make a usable vehicle lights toggle but its still under progress.

FINDINGS: Apparently the helis (and depending on how exquisite you are also the planes) need a separated preset for Nightvision, there is an annoying "tilt" effect or kinda stroboscopic effect in the UI of them but this doesn't happen for other land vehicles or for some other air vehicles (such as the AC-130 or others)

NOTES: Might be a good idea to remove the time enable stuff in the settings of DU, since this is fully controled by VEM now, at max, DU should call VEM modules to register presets, but not control anything related to time or so.